requirements.txt
groq
python-dotenv
pyspellchecker
colorama